refactor courier test & add report allure
