--- src/test/java/RestAssuredGetAllureTest.java
@@ -1,51 +1,0 @@
-import io.restassured.RestAssured;
-import io.restassured.response.Response;
-import org.junit.Before;
-import org.junit.Test;
-import io.qameta.allure.junit4.DisplayName; // импорт DisplayName
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.Matchers.equalTo;
-
-public class RestAssuredGetAllureTest {
-
-    String bearerToken = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2MjZhZTBiODA0NmM5ZDAwM2QxNmYwYTEiLCJpYXQiOjE2NTcyNzQ0MjMsImV4cCI6MTY1Nzg3OTIyM30.jQlwh8NTu3dCmmX_VON34BboWTfaT-LwNvy8tC1Jans";
-
-    @Before
-    public void setUp() {
-        RestAssured.baseURI= "https://qa-mesto.praktikum-services.ru";
-    }
-
-    @Test
-    @DisplayName("Check status code of /users/me") // имя теста
-    public void getMyInfoStatusCode() {
-        given()
-                .auth().oauth2(bearerToken)
-                .get("/api/users/me")
-                .then().statusCode(200);
-    }
-
-    @Test
-    @DisplayName("Check user name") // имя теста
-    public void checkUserName() {
-        given()
-                .auth().oauth2(bearerToken)
-                .get("/api/users/me")
-                .then().assertThat().body("data.name",equalTo("Василий Васильев"));
-    }
-
-    @Test
-    @DisplayName("Check user name and print response body") // имя теста
-    public void checkUserNameAndPrintResponseBody() {
-
-        Response response = given().auth().oauth2(bearerToken).get("/api/users/me");
-        // отправили запрос и сохранили ответ в переменную response - экземпляр класса Response
-
-        response.then().assertThat().body("data.name",equalTo("Василий Васильев"));
-        // проверили, что в теле ответа ключу name соответствует нужное имя пользователя
-
-        System.out.println(response.body().asString()); // вывели тело ответа на экран
-
-    }
-
-}
